Organize code base
